patch: Fix discover page breaking when removing a query
When the user deletes their query without submitting, the entire
search interface vanishes. When the user submits the new query
in the same scenario (by pressing the 'clear' button)
the "no connections" interface would flash in while the new set
of results load.

This fixes both issues by checking whether or not the user has
*any* connections as opposed to whether or not there is a search
term or results